move ui tabs from Graphs to main app
split Graphs into PlotTab and LiveTab.
the live tab is intended to show live data,
whereas the plot tab is intended for analyzing/browsing the data.
the trajectory tab is intended to be a 3d visualization of the flight
path and stages.
the network tab is intended for analyzing MAVLink data and would have a
log of every packet sent/received.